Django-signals_orm-0x04/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-filter==24.3
drf-nested-routers==0.94.1
djangorestframework-simplejwt==5.5.0
